Move G2 context to the `ui` folder
