--- packages/components/src/__next/context/index.js
@@ -1,2 +1,0 @@
-export { ComponentSystemProvider as __unstableComponentSystemProvider } from './component-system-provider';
-export { withNext as __unstableWithNext } from './with-next';
